# Don't allocate an extra pointer in `mimirpb.ReuseSlice()`