#### What this PR does

We can avoid allocating an extra pointer just to put something into a pool if we store the pointer we got from the first pool in a second pool. Used the implementation of a pool from Prometheus for this.

```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/util/push
cpu: 11th Gen Intel(R) Core(TM) i7-11700K @ 3.60GHz
               │   old.txt   │            new.txt            │
               │   sec/op    │   sec/op     vs base          │
PushHandler-16   1.635µ ± 8%   1.608µ ± 4%  ~ (p=0.289 n=10)

               │  old.txt   │              new.txt              │
               │    B/op    │    B/op     vs base               │
PushHandler-16   659.0 ± 0%   633.0 ± 0%  -3.95% (p=0.000 n=10)

               │  old.txt   │              new.txt              │
               │ allocs/op  │ allocs/op   vs base               │
PushHandler-16   17.00 ± 0%   16.00 ± 0%  -5.88% (p=0.000 n=10)
```

#### Which issue(s) this PR fixes or relates to

None, was just passing by.

#### Checklist

- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`