@colega 2023-03-27
Also fixed the `BenchmarkPushHandler` which was not parsing the request at all, and after that it shows that one allocation saved. The cpu time difference gets diluted by the amount of time spent on the rest of the things.

```
$ benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/util/push
cpu: 11th Gen Intel(R) Core(TM) i7-11700K @ 3.60GHz
               │   old.txt   │            new.txt            │
               │   sec/op    │   sec/op     vs base          │
PushHandler-16   1.552µ ± 5%   1.496µ ± 5%  ~ (p=0.072 n=10)

               │  old.txt   │              new.txt              │
               │    B/op    │    B/op     vs base               │
PushHandler-16   658.0 ± 0%   634.0 ± 1%  -3.65% (p=0.000 n=10)

               │  old.txt   │              new.txt              │
               │ allocs/op  │ allocs/op   vs base               │
PushHandler-16   17.00 ± 0%   16.00 ± 0%  -5.88% (p=0.000 n=10)
```