@colega 2023-03-28
Lets wait until https://github.com/prometheus/prometheus/pull/12189 lands and use that instead of ad-hoc code.