@pracucci 2023-03-29
> Lets wait until [prometheus/prometheus#12189](https://github.com/prometheus/prometheus/pull/12189) lands and use that instead of ad-hoc code.

I looked at it, and I love it. Smart approach! Looking forward to get that PR merged and then used here.