@colega 2023-04-17
I rebased the branch using zeropool and updated the PR description, @pracucci PTAL.