@colega 2023-04-17
There's no changelog entry for this (it's a minor implementation change, I don't think it's worth it) so going to merge it once the CI passes.